google-sheets/pyproject.toml
[project]
name = "google-sheets-skill"
version = "0.1.0"
description = "Codex skill helpers for Google Sheets"
requires-python = ">=3.10"
dependencies = [
  "google-api-python-client>=2.170.0",
  "google-auth>=2.40.0",
  "google-auth-oauthlib>=1.2.0",
  "regex>=2024.11.6",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["google_sheets_skill*"]
